There is still a lot of tweaking and moving logic around to better fit into Laravel system.
I moved validation and authorization from the request object down to the models because that will let me deal with it closer to where it is important, I don't have to expose the inner working of the model and I give each model full control over how it manages authorization/request data.

I also broke with Laravel convention of storing custom service provider classes in the Providers folder, because once I'm done implementing the MVSC system here I want to easily be able to reuse it in other apps.